Major refactor on how information is passed to the shader for lighting.
- Created several helper classes which abstract light information.
-- OpenGLPointLightData is the finalized data type to pass to OpenGL.
-- OpenGLAbstractLightGroup is for different types of lights.
- OpenGLMesh was altered to bind the VAO instead of having to get VAO.
- OpenGLInstanceGroup uses the new OpenGLMesh bindings now.
- OpenGLMesh is reference counted.
- KSharedPointer is a method of reference counting information.
- KVertex is now Pos+Normal, KDebugVertex is Pos+Color
- KMath holds translation functions from Q* types to glm. (temporary)
- Karma includes were moved to another subfolder for ease of include.
- PointLight.vert no longer takes ambience into account. (unused)

+++ Karma/kstaticgeometry.cpp
@@ -254,10 +254,10 @@ void KStaticGeometryPrivate::buildBottomUp(TerminationPred pred)
 
   int leafCount;
   int depthEstimate;
-  int leafSize = numTriangles;
+  int leafSize = static_cast<int>(numTriangles);
   do
   {
-    leafCount = numTriangles / leafSize;
+    leafCount = static_cast<int>(numTriangles / leafSize);
     depthEstimate = std::ceil(std::log(leafSize) / Karma::Log2);
     leafSize /= 2;
   }
